+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Parse left and right notes in tuc
 - Add tuc part
     - `ReadEvent` from stream
+- Show better error message for undefined component field
 
 ## 0.1.0 - 2020-08-12
 - Initial implementation